profiles/gap: Change len parameter of name2utf8() to uint16_t
When device name is read its length is of type uint16_t, change the type
of length parameter of name2utf8() to uint16_t to avoid unnecessary
truncation.

+++ profiles/gap/gas.c
@@ -73,7 +73,7 @@ static int cmp_device(gconstpointer a, gconstpointer b)
 	return gas->device == device ? 0 : -1;
 }
 
-static char *name2utf8(const uint8_t *name, uint8_t len)
+static char *name2utf8(const uint8_t *name, uint16_t len)
 {
 	char utf8_name[HCI_MAX_NAME_LENGTH + 2];
 	int i;
